+++ src/test/java/render/text/TCBFontTest.java
@@ -1,12 +1,10 @@
-package TCBTest;
+package render.text;
 
 import TheCellBeyond.internal.ResourceID;
 import TheCellBeyond.internal.ResourceStatus;
 import TheCellBeyond.internal.ResourceStatusCallback;
 import TheCellBeyond.internal.ResourceStatusListener;
 import org.junit.jupiter.api.Test;
-import render.text.GlyphRange;
-import render.text.TCBFont;
 import utility.AssetManager;
 import utility.Settings;
